Moved git_suffix in baselib
