+++ openquake/baselib/general.py
@@ -336,6 +336,22 @@ def writetmp(content=None, dir=None, prefix="tmp", suffix="tmp"):
     return path
 
 
+def git_suffix(fname):
+    """
+    :returns: `<short git hash>` if Git repository found
+    """
+    try:
+        gh = subprocess.check_output(
+            ['git', 'rev-parse', '--short', 'HEAD'],
+            stderr=open(os.devnull, 'w'), cwd=os.path.dirname(fname)).strip()
+        gh = "-git" + gh if gh else ''
+        return gh
+    except:
+        # trapping everything on purpose; git may not be installed or it
+        # may not work properly
+        return ''
+
+
 def run_in_process(code, *args):
     """
     Run in an external process the given Python code and return the
