+++ openquake/hazardlib/__init__.py
@@ -17,11 +17,11 @@
 hazardlib stands for Hazard Library.
 """
 
+from openquake.baselib.general import git_suffix
 from openquake.hazardlib import (
     calc, geo, gsim, mfd, scalerel, source, const, correlation, imt, pmf, site,
-    tom, general, near_fault
-)
+    tom, near_fault)
 
 # the version is managed by packager.sh with a sed
 __version__ = '0.15.0'
-__version__ += general.git_suffix(__file__)
+__version__ += git_suffix(__file__)
